Fix logic problem
There is a flaw in the current logic at the end of run_protocols() when testing a TLSv1.3-only server using an $OPENSSL that does not support TLSv1.3. If $OSSL_SHORTCUT is true and /usr/bin/openssl either does not exist or is not executable, then testssl.sh will continue on using /usr/bin/openssl as $OPENSSL even if /usr/bin/openssl does not support TLSv1.3.

This commit fixes the issue by reorganizing the logic.